requirements.txt
Flask
Flask-SQLAlchemy
flask_login
pymysql
requests
bs4
gunicorn
